update weight limit to be float and add unit test for create drone repository

+++ repository/entity/drone.go
@@ -6,7 +6,7 @@ type Drone struct {
 	gorm.Model
 	SerialNumber    string         `json:"serial_number" gorm:"size:100;unique"`
 	DroneModel      DroneModelEnum `json:"drone_model"`
-	WeightLimit     uint           `json:"weight_limit" gorm:"check:weight_limit <= 500"`
+	WeightLimit     float64        `json:"weight_limit" gorm:"check:weight_limit <= 500"`
 	BatteryCapacity uint           `json:"battery_capacity" gorm:"check:battery_capacity <= 100"`
 	State           DroneStateEnum `json:"state"`
 }
@@ -31,7 +31,7 @@ const (
 type DroneRequest struct {
 	SerialNumber    string         `json:"serial_number"`
 	DroneModel      DroneModelEnum `json:"drone_model"`
-	WeightLimit     uint           `json:"weight_limit"`
+	WeightLimit     float64        `json:"weight_limit"`
 	BatteryCapacity uint           `json:"battery_capacity"`
 	State           DroneStateEnum `json:"state"`
 }
